feat: Axelar Bridge as EVMs default
